requirement.txt
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
audioop-lts==0.2.1
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cmake==4.0.2
colorama==0.4.6
comm==0.2.2
comtypes==1.4.11
contourpy==1.3.1
cryptography==45.0.3
curl_cffi==0.11.1
cycler==0.12.1
dashscope==1.23.3
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastcore==1.7.29
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.56.0
fpdf==1.7.2
fpdf2==2.8.3
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.31.0
gradio_client==1.10.1
greenlet==3.2.2
groovy==0.1.2
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
idna==3.10
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
Js2Py==0.74
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==1.1.11
keras==3.10.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
langsmith==0.3.42
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.1.1
ml_dtypes==0.5.1
mlxtend==0.23.4
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
multitasking==0.0.11
mutagen==1.47.0
mypy_extensions==1.1.0
mysql==0.0.3
mysqlclient==2.2.7
namex==0.1.0
narwhals==1.28.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.3
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
optree==0.16.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peewee==3.18.1
pillow==11.1.0
pipwin==0.5.2
platformdirs==4.3.6
plotly==6.0.0
plotly-express==0.4.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
pyjsparser==2.7.1
PyMySQL==1.1.1
pyparsing==3.2.2
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
pyspellchecker==0.8.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.1
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
ruff==0.11.11
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
secure-smtplib==0.1.1
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.2
statsmodels==0.14.4
stqdm==0.0.5
streamlit==1.42.2
streamlit-jupyter==0.2.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
typer==0.15.4
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
ultralytics==8.3.135
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
uv==0.7.9
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==3.6.10
wikipedia==1.4.0
wordcloud==1.9.4
xgboost==3.0.0
yarl==1.20.0
yfinance==0.2.61
yt-dlp==2025.6.9
zstandard==0.23.0
